IMPORTANT: MASSIVE REPOSITORY REORG - READ ME!
IF YOUR REPO IS INITIALIZED ON libaro42, YOU _NEED_ TO repo init jb42.

ALL MERGED COMMITS IN THE FOLLOWING PROJECTS HAVE BEEN REBASED ONTO THE LINARO BASE
UNMERGED COMMITS IN GERRIT ON THE FOLLOWING PROJECTS WILL NEED TO BE REBASED on vanir/jb42!!

    build
    bionic
    dalvik
    device/samsung/tuna
    frameworks/base
    frameworks/native
    hardware/libhardware
    system/core
    packages/inputmethods/LatinIME
    external/busybox
    hardware/ti/omap4xxx
    manifest

PROJECTS THAT WERE NOT COMPLETELY MERGED/REBASED ON LINARO CHANGES ARE TAGGED "TODO" IN default.xml ON JB42.