change to this kernel so anything msm8960 can use F2FS
i have enabled f2fs in this kernel for all msm8960 devices. this will be
updated with cm and f2fs source on a regular basis and is
serperated from all other kernel projects so it is not going to get deleted and break shit. it has beenn tested on all devices and will works on all
of them

Change-Id: I2a28b858e65ddc6ddbb20dedc438f47f0415093a